feat: add simplified useAnimatedVariantColor hook for color interpolation
